game function created
